+++ src/index.js
@@ -25,6 +25,7 @@ import DepthCloudObject from './viz/DepthCloud';
 import Viewer2d from './viewers/2d';
 import Viewer3d from './viewers/3d';
 import TfViewer from './viewers/Tf';
+import InteractiveMarkers from './viz/InteractiveMarkers';
 
 export default {
   CollisionObject,
@@ -53,4 +54,6 @@ export default {
   Viewer3d,
   TfViewer,
   DepthCloudObject,
+
+  InteractiveMarkers,
 };